LayoutTests/platform/mac-wk1/fast/events/unload-showModalDialog-expected.txt

main frame - has 1 onunload handler(s)
CONSOLE MESSAGE: Window 'showModalDialog' function is deprecated and will be removed soon.
CONSOLE MESSAGE: showModalDialog() is deprecated and will be removed. Please use the <dialog> element instead.
CONSOLE MESSAGE: Use of window.showModalDialog is not allowed while unloading a page.
This test verifies that prompts are forbidden when leaving a webpage.

Prompts include alert, confirm, prompt, and showModalDialog.

If there is no prompt, the test passed.
